+++ tests/fixtures/cases/outputs/generate.js
@@ -5,7 +5,7 @@ const type = process.argv[2];
 const inWorkspace = process.argv.includes('--workspace');
 
 function createFile(file, content) {
-	const filePath = path.join(inWorkspace ? process.env.MOON_WORKSPACE_ROOT : __dirname, file);
+	const filePath = path.join(inWorkspace ? process.env.MOON_WORKSPACE_ROOT : __dirname, type, file);
 
 	fs.mkdirSync(path.dirname(filePath), { recursive: true });
 	fs.writeFileSync(filePath, content ?? String(Date.now()), 'utf8');
@@ -13,23 +13,23 @@ function createFile(file, content) {
 
 switch (type) {
 	case 'single-file':
-		createFile('lib/one.js');
+		createFile('one.js');
 		break;
 	case 'single-folder':
 	case 'multiple-files':
-		createFile('lib/one.js');
-		createFile('lib/two.js');
+		createFile('one.js');
+		createFile('two.js');
 		break;
 	case 'multiple-folders':
 	case 'both':
-		createFile('lib/one.js');
-		createFile('esm/two.js');
+		createFile('a/one.js');
+		createFile('b/two.js');
 		break;
 	case 'multiple-types':
-		createFile('build/one.js');
-		createFile('build/two.js');
-		createFile('build/styles.css');
-		createFile('build/image.png');
+		createFile('one.js');
+		createFile('two.js');
+		createFile('styles.css');
+		createFile('image.png');
 		break;
 	case 'custom':
 		createFile(process.argv[3], 'fixed content');